add RightsManager, InternalUsersAuthenticationProvider, @REST.Method not restricted to a POST, get session asynchronously, bundle can scan packages to search services automatically

+++ web/src/main/java/net/lecousin/framework/web/services/WebService.java
@@ -61,7 +61,7 @@ public interface WebService {
 		/** Right name. */
 		String name();
 		/** Right value. */
-		int value();
+		long value();
 	}
 
 	/** Specify that a service or method of a service needs rights to be executed. */
